Code cleanup, moved subclasses inside module.hpp to their own files
